TidyUp: 2. BikeController: "return ResponseEntity.ok(BikeDto.fromEntity(bikeService.addNewBike(bikeDto.toEntity())));" looks a bit shit.

+++ src/main/java/com/bike/service/BikeService.java
@@ -45,6 +45,9 @@ public class BikeService {
   }
 
   public void deleteBike(UUID bikeId) {
+    if (!bikeRepository.existsById(bikeId)) {
+      throw new EntityNotFoundException("Unable to find Bike with id " + bikeId);
+    }
     bikeRepository.deleteById(bikeId);
   }
